+++ app/Http/Controllers/StreetController.php
@@ -2,6 +2,8 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\City;
+use App\Models\Neighborhood;
 use Illuminate\Http\Request;
 use App\Models\Street;
 
@@ -12,10 +14,10 @@ class StreetController extends Controller
      */
     public function index()
     {
-        $street = Street::all();
+        $street = City::with('neighborhood.street')->get();
 
-        if($street === null){
-            return response()->json(['message'=> 'Nenhuma rua encontrado!'], 404);
+        if ($street === null) {
+            return response()->json(['message' => 'Nenhuma rua encontrado!'], 404);
         }
 
         return response()->json($street, 200);
@@ -26,7 +28,9 @@ class StreetController extends Controller
      */
     public function create()
     {
-        //
+        $neighborhood = Neighborhood::select('id', 'name')->get();
+
+        return response()->json($neighborhood, 200);
     }
 
     /**
@@ -35,8 +39,22 @@ class StreetController extends Controller
     public function store(Request $request)
     {
         $street = Street::create($request->all());
-        
+
         return response()->json($street, 201);
+    }
+
+    /**
+     * Show the form for editing the specified resource.
+     */
+    public function edit(string $street_id)
+    {
+        $street = Street::find($street_id);
+
+        if ($street === null) {
+            return response()->json(['message' => 'Rua não encontrada!'], 404);
+        }
+
+        return response()->json($street, 200);
     }
 
     /**
@@ -44,8 +62,8 @@ class StreetController extends Controller
      */
     public function update(Request $request, Street $street)
     {
-        if($street === null){
-            return response()->json(['message'=> 'Rua não encontrada!'], 404);
+        if ($street === null) {
+            return response()->json(['message' => 'Rua não encontrada!'], 404);
         }
         $street->fill($request->all());
         $street->save();
@@ -54,16 +72,30 @@ class StreetController extends Controller
     }
 
     /**
+     * Display the specified resource.
+     */
+    public function show(string $street_id)
+    {
+        $street = Street::find($street_id);
+
+        if ($street === null) {
+            return response()->json(['message' => 'Rua não encontrada!'], 404);
+        }
+
+        return response()->json($street, 200);
+    }
+
+    /**
      * Remove the specified resource from storage.
      */
     public function destroy(string $street_id)
     {
-        // $street_id = Street::whereId($street_id)->first();
+        $street = Street::whereId($street_id)->first();
 
-        // if($street_id === null){
-        //     return response()->json(['message'=> ' Bairro não encontrado!'], 404);
-        // }
-        
+        if ($street === null) {
+            return response()->json(['message' => ' Rua não encontrada!'], 404);
+        }
+
         Street::destroy($street_id);
         return response()->noContent();
     }
